move canvas to separate component so code can be exported
