+++ app/(index)/DescriptionSection.tsx
@@ -7,7 +7,7 @@ import { useFrame } from "@react-three/fiber";
 
 import { getViewportHeightAtDepth } from "@/app/utils";
 
-import { CAMERA_SETTINGS } from "./page";
+import { CAMERA_SETTINGS } from "./Canvas";
 import { SCENE_MARGIN_X_TO_VIEWPORT_RATIO } from "./ScrollableScene";
 
 const DescriptionSection = () => {
